Deduplicate contract source response
